Fix PYENV_VERSION usage in python.sh
Unfortunately, pyenv uses `PYENV_VERSION` variable to choose the
version of Python. In this script, `PYENV_VERSION` is used to choose
the version of pyenv itself. If this variable is exported (for
example, in project settings), pyenv gets confused and fails as it
cannot find non-existing versions of Python.

Example. The following variables are exported:

    PYENV_VERSION=v1.0.10
    PYTHON_VERSION=3.6.1

When python.sh is executed, `python --version` command fails:

    pyenv: version `v1.0.10' is not installed (set by PYENV_VERSION environment variable)

The simplest solution (this commit) is to unset the `PYENV_VERSION`
variable before actually using pyenv. Alternatively, another variable
name could be used to avoid confusion, but other names look awkward
(`PYENV_TOOL_VERSION`?).